third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/flow-content-0/search-styles-iso-8859-8-expected.txt

This is a testharness.js-based test.
PASS Verify document.characterSet is ISO-8859-8
FAIL <search> - display assert_equals: expected "block" but got "inline"
PASS <search> - margin-top
PASS <search> - margin-right
PASS <search> - margin-bottom
PASS <search> - margin-left
PASS <search> - padding-top
PASS <search> - padding-right
PASS <search> - padding-bottom
PASS <search> - padding-left
FAIL <search> - unicode-bidi assert_equals: expected "bidi-override" but got "normal"
Harness: the test ran to completion.
